perf: improve npm-installer.sh

+++ command/npm-installer.sh
@@ -4,35 +4,46 @@ set -e
 
 TMP_DIR="/var/www/tmp"
 
-if [ $1 == '--help' ]; then
-  echo "NPM Installer script v1.0.0"
-  echo "usage: bash bin/npm-installer.sh [install|update|move_tmp|move_current]"
-  exit 0
-fi
+display_help() {
+	echo "NPM Installer script v1.1.0"
+	echo "usage: bash $(basename $0) [install|update|update_all|move_tmp|move_current]"
+	exit 0
+}
 
 if [ $# -ne 1 ]; then
-  echo 'Error: Required One arguments'
-  exit 1
+	echo 'Error: Required One arguments'
+	display_help
+	exit 1
+fi
+
+if [ $1 == '--help' ]; then
+	display_help
 fi
 
 CURRENT_DIR=$(pwd)
 echo "current dir: ${CURRENT_DIR}"
 
 if [ -z "$CURRENT_DIR" ]; then
-  echo 'Error: current dir not found'
-  exit 1
+	echo 'Error: current dir not found'
+	exit 1
 fi
 
 echo "temp dir: ${TMP_DIR}"
 
 if [ ! -d "${TMP_DIR}" ]; then
-  echo 'Error: tmp dir not found'
-  exit 1
+	echo 'Error: tmp dir not found'
+	exit 1
 fi
 
 npm_install() {
 	echo "[Info] npm install";
 	npm install
+	npm audit fix
+}
+
+npm_update() {
+	echo "[Info] npm update";
+	npm update
 	npm audit fix
 }
 
@@ -44,9 +55,17 @@ update_package_json() {
 
 move_to_tmp_dir() {
 	echo "[Info] Moving package.json, package-lock.json and node_modules to ${TMP_DIR}";
-	mv package.json ${TMP_DIR}
-	mv package-lock.json ${TMP_DIR}
-	mv node_modules ${TMP_DIR}
+
+	if [ -f package.json ]; then
+		mv package.json ${TMP_DIR}
+	fi
+	if [ -f package-lock.json ]; then
+		mv package-lock.json ${TMP_DIR}
+	fi
+	if [ -d node_modules ]; then
+		mv node_modules ${TMP_DIR}
+	fi
+
 	cd ${TMP_DIR}
 	echo "[Info] move working dir to ${TMP_DIR}";
 }
@@ -55,9 +74,16 @@ move_to_current_dir() {
 	cd ${CURRENT_DIR}
 	echo "[Info] move working dir to ${CURRENT_DIR}";
 	echo "[Info] Moving package.json, package-lock.json and node_modules to current dir";
-	mv ${TMP_DIR}/package.json ./
-	mv ${TMP_DIR}/package-lock.json ./
-	mv ${TMP_DIR}/node_modules ./
+
+	if [ -f ${TMP_DIR}/package.json ]; then
+		mv ${TMP_DIR}/package.json ./
+	fi
+	if [ -f ${TMP_DIR}/package-lock.json ]; then
+		mv ${TMP_DIR}/package-lock.json ./
+	fi
+	if [ -d ${TMP_DIR}/node_modules ]; then
+		mv ${TMP_DIR}/node_modules ./
+	fi
 }
 
 if [ $1 = 'install' ]; then
@@ -66,8 +92,12 @@ if [ $1 = 'install' ]; then
 	move_to_current_dir
 elif [ $1 = 'update' ]; then
 	move_to_tmp_dir
+	npm_update
+	move_to_current_dir
+elif [ $1 = 'update_all' ]; then
+	move_to_tmp_dir
 	update_package_json
-	npm_install
+	npm_update
 	move_to_current_dir
 elif [ $1 = 'move_tmp' ]; then
 	move_to_tmp_dir
